UtilProperties bug fix and enhancements.
This commit fixes the bug reported in https://issues.apache.org/jira/browse/OFBIZ-1488 and it includes a few of the improvements mentioned in https://issues.apache.org/jira/browse/OFBIZ-1485. This commit also lays the groundwork for improved properties file support in the future.

This commit contains a new xsd file - it needs to be included on the apache ofbiz website. XML properties files will generate console error messages until it is.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/framework/webtools@603024 13f79535-47bb-0310-9956-ffa450edef68