A code cleanup/reorganization in the service engine.
The LocalDispatcher is supposed to represent the state of a single application, but that state was split between LocalDispatcher and DispatchContext. In addition, both classes shared a lot of fields.

This commit moves application state to LocalDispatcher and DispatchContext becomes a lightweight container for objects service implementations need to use.

In addition, some minor thread-safety issues were fixed.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/framework/webtools@1622256 13f79535-47bb-0310-9956-ffa450edef68