Revert rev 1622256 to fix a service definition caching problem - https://issues.apache.org/jira/browse/OFBIZ-6060.
I'm reverting this instead of fixing it because the original commit was a first step in an effort to eliminate ThreadLocal variables in the framework (so we can make better use of multi-threading). But since then I have come to the conclusion that it will not be possible without a framework redesign.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/framework/webtools@1657714 13f79535-47bb-0310-9956-ffa450edef68