A patch from Gil Portenseigne for "Add a mean to untie geo associations" https://issues.apache.org/jira/browse/OFBIZ-6528
In the Webtools we can asociate Geos but we have yet no means to untie them

This patch adds a new grid in EditGeoScreen to see/delete GeoAssoc and CommonEntityLabels default-resource-name on GeoAssocType for translation purpose.

jleroux: On Pierre's Smits suggestion Gil also added a clickable link to open the associated Geo and a column to show the geoTypeId of the associated Geo.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/framework/webtools@1687259 13f79535-47bb-0310-9956-ffa450edef68